Move GlobalOnlyResolver into coded_tools/utils
GlobalOnlyResolver is a pure helper (an aiohttp AbstractResolver), not a
registered CodedTool, and is used only by SafeFetch. Relocating it alongside
safe_fetch.py and pdf_utils.py puts the complete SSRF-hardened fetch unit in
one package. Updates the two importers (safe_fetch.py and the resolver's test);
no config/registry references existed.
